review update

+++ backend/app/api/reviews.py
@@ -23,6 +23,8 @@ def review_by_id(review_id):
     if request.method == 'GET':
         return jsonify({"review": review.to_dict()}), 200
     elif request.method == 'PUT':
+        if review.user_id != get_jwt_identity():
+            return jsonify({"error": "You are not allowed to edit this review"}), 403
         allowed_keys = ['score', 'comment']
         if not all(key in allowed_keys for key in data.keys()):
             return jsonify({"error": "Invalid input"}), 400
@@ -31,8 +33,11 @@ def review_by_id(review_id):
         db.session.commit()
         return jsonify({"review": review.to_dict()}), 200
     elif request.method == 'DELETE':
+        if review.user_id != get_jwt_identity():
+            return jsonify({"error": "You are not allowed to delete this review"}), 403
         db.session.delete(review)
         db.session.commit()
+        return jsonify({"message": "Review delited"}), 200
 
 
 @bp.route('/reviews/<int:dish_id>', methods=['POST'])
